requirements.txt
# Core dependencies
requests>=2.31.0
rich>=13.0.0
pyyaml>=6.0

# Development/testing dependencies (optional)
pytest>=7.0.0

# Note: The following are Python standard library modules (no install needed):
# - json, os, sys, time, pathlib, datetime, typing, dataclasses
# - logging, argparse, subprocess, importlib, inspect